DailyWork/Demo/FlaskProject/requirements.txt
Flask
mysql-connector
Flask-MySQLdb